+++ tests/color_tests.rs
@@ -1,4 +1,5 @@
 use fast_image_resize as fr;
+use fast_image_resize::pixels::*;
 use testing::nonzero;
 
 mod gamma_tests {
@@ -9,24 +10,24 @@ mod gamma_tests {
         let mapper = fr::create_gamma_22_mapper();
         let buffer: Vec<u8> = (0u8..=255).collect();
         let src_image =
-            fr::Image::from_vec_u8(nonzero(16), nonzero(16), buffer, fr::PixelType::U8).unwrap();
-        let src_checksum = testing::image_checksum::<1>(src_image.buffer());
+            fr::Image::from_vec_u8(nonzero(16), nonzero(16), buffer, PixelType::U8).unwrap();
+        let src_checksum = testing::image_checksum::<U8, 1>(&src_image);
         assert_eq!(src_checksum, [32640]);
 
         // into U8
-        let mut dst_image = fr::Image::new(nonzero(16), nonzero(16), fr::PixelType::U8);
+        let mut dst_image = fr::Image::new(nonzero(16), nonzero(16), PixelType::U8);
         mapper
             .forward_map(&src_image.view(), &mut dst_image.view_mut())
             .unwrap();
-        let dst_checksum = testing::image_checksum::<1>(dst_image.buffer());
+        let dst_checksum = testing::image_checksum::<U8, 1>(&dst_image);
         assert_eq!(dst_checksum, [20443]);
 
         // into U16
-        let mut dst_image = fr::Image::new(nonzero(16), nonzero(16), fr::PixelType::U16);
+        let mut dst_image = fr::Image::new(nonzero(16), nonzero(16), PixelType::U16);
         mapper
             .forward_map(&src_image.view(), &mut dst_image.view_mut())
             .unwrap();
-        let dst_checksum = testing::image_u16_checksum::<1>(dst_image.buffer());
+        let dst_checksum = testing::image_checksum::<U16, 1>(&dst_image);
         assert_eq!(dst_checksum, [5255141]);
     }
 
@@ -35,13 +36,13 @@ mod gamma_tests {
         let mapper = fr::create_gamma_22_mapper();
         let buffer: Vec<u8> = (0u8..=255).collect();
         let src_image =
-            fr::Image::from_vec_u8(nonzero(16), nonzero(16), buffer, fr::PixelType::U8).unwrap();
+            fr::Image::from_vec_u8(nonzero(16), nonzero(16), buffer, PixelType::U8).unwrap();
 
-        let mut dst_image = fr::Image::new(nonzero(16), nonzero(1), fr::PixelType::U8);
+        let mut dst_image = fr::Image::new(nonzero(16), nonzero(1), PixelType::U8);
         let result = mapper.forward_map(&src_image.view(), &mut dst_image.view_mut());
         assert!(matches!(result, Err(fr::MappingError::DifferentDimensions)));
 
-        let mut dst_image = fr::Image::new(nonzero(16), nonzero(16), fr::PixelType::U8x2);
+        let mut dst_image = fr::Image::new(nonzero(16), nonzero(16), PixelType::U8x2);
         let result = mapper.forward_map(&src_image.view(), &mut dst_image.view_mut());
         assert!(matches!(
             result,
@@ -54,16 +55,16 @@ mod gamma_tests {
         let mapper = fr::create_gamma_22_mapper();
         let buffer: Vec<u8> = (0u8..=255).collect();
         let src_image =
-            fr::Image::from_vec_u8(nonzero(16), nonzero(16), buffer, fr::PixelType::U8).unwrap();
-        let mut dst_image = fr::Image::new(nonzero(16), nonzero(16), fr::PixelType::U8);
+            fr::Image::from_vec_u8(nonzero(16), nonzero(16), buffer, PixelType::U8).unwrap();
+        let mut dst_image = fr::Image::new(nonzero(16), nonzero(16), PixelType::U8);
 
         mapper
             .backward_map(&src_image.view(), &mut dst_image.view_mut())
             .unwrap();
 
-        let src_checksum = testing::image_checksum::<1>(src_image.buffer());
+        let src_checksum = testing::image_checksum::<U8, 1>(&src_image);
         assert_eq!(src_checksum, [32640]);
-        let dst_checksum = testing::image_checksum::<1>(dst_image.buffer());
+        let dst_checksum = testing::image_checksum::<U8, 1>(&dst_image);
         assert_eq!(dst_checksum, [44824]);
     }
 
@@ -72,13 +73,13 @@ mod gamma_tests {
         let mapper = fr::create_gamma_22_mapper();
         let buffer: Vec<u8> = (0u8..=255).collect();
         let src_image =
-            fr::Image::from_vec_u8(nonzero(16), nonzero(16), buffer, fr::PixelType::U8).unwrap();
+            fr::Image::from_vec_u8(nonzero(16), nonzero(16), buffer, PixelType::U8).unwrap();
 
-        let mut dst_image = fr::Image::new(nonzero(16), nonzero(1), fr::PixelType::U8);
+        let mut dst_image = fr::Image::new(nonzero(16), nonzero(1), PixelType::U8);
         let result = mapper.backward_map(&src_image.view(), &mut dst_image.view_mut());
         assert!(matches!(result, Err(fr::MappingError::DifferentDimensions)));
 
-        let mut dst_image = fr::Image::new(nonzero(16), nonzero(16), fr::PixelType::U8x2);
+        let mut dst_image = fr::Image::new(nonzero(16), nonzero(16), PixelType::U8x2);
         let result = mapper.backward_map(&src_image.view(), &mut dst_image.view_mut());
         assert!(matches!(
             result,
@@ -95,16 +96,16 @@ mod srgb_tests {
         let mapper = fr::create_srgb_mapper();
         let buffer: Vec<u8> = (0u8..=255).flat_map(|v| [v, v, v]).collect();
         let src_image =
-            fr::Image::from_vec_u8(nonzero(16), nonzero(16), buffer, fr::PixelType::U8x3).unwrap();
-        let src_checksum = testing::image_checksum::<3>(src_image.buffer());
+            fr::Image::from_vec_u8(nonzero(16), nonzero(16), buffer, PixelType::U8x3).unwrap();
+        let src_checksum = testing::image_checksum::<U8x3, 3>(&src_image);
         assert_eq!(src_checksum, [32640, 32640, 32640]);
 
-        let mut dst_image = fr::Image::new(nonzero(16), nonzero(16), fr::PixelType::U8x3);
+        let mut dst_image = fr::Image::new(nonzero(16), nonzero(16), PixelType::U8x3);
         mapper
             .forward_map(&src_image.view(), &mut dst_image.view_mut())
             .unwrap();
 
-        let dst_checksum = testing::image_checksum::<3>(dst_image.buffer());
+        let dst_checksum = testing::image_checksum::<U8x3, 3>(&dst_image);
         assert_eq!(dst_checksum, [20304, 20304, 20304]);
     }
 
@@ -113,16 +114,16 @@ mod srgb_tests {
         let mapper = fr::create_srgb_mapper();
         let buffer: Vec<u8> = (0u8..=255).flat_map(|v| [v, v, v, 255]).collect();
         let src_image =
-            fr::Image::from_vec_u8(nonzero(16), nonzero(16), buffer, fr::PixelType::U8x4).unwrap();
-        let src_checksum = testing::image_checksum::<4>(src_image.buffer());
+            fr::Image::from_vec_u8(nonzero(16), nonzero(16), buffer, PixelType::U8x4).unwrap();
+        let src_checksum = testing::image_checksum::<U8x4, 4>(&src_image);
         assert_eq!(src_checksum, [32640, 32640, 32640, 65280]);
 
-        let mut dst_image = fr::Image::new(nonzero(16), nonzero(16), fr::PixelType::U8x4);
+        let mut dst_image = fr::Image::new(nonzero(16), nonzero(16), PixelType::U8x4);
         mapper
             .forward_map(&src_image.view(), &mut dst_image.view_mut())
             .unwrap();
 
-        let dst_checksum = testing::image_checksum::<4>(dst_image.buffer());
+        let dst_checksum = testing::image_checksum::<U8x4, 4>(&dst_image);
         assert_eq!(dst_checksum, [20304, 20304, 20304, 65280]);
     }
 
@@ -131,13 +132,13 @@ mod srgb_tests {
         let mapper = fr::create_srgb_mapper();
         let buffer: Vec<u8> = (0u8..=255).flat_map(|v| [v, v, v]).collect();
         let src_image =
-            fr::Image::from_vec_u8(nonzero(16), nonzero(16), buffer, fr::PixelType::U8x3).unwrap();
+            fr::Image::from_vec_u8(nonzero(16), nonzero(16), buffer, PixelType::U8x3).unwrap();
 
-        let mut dst_image = fr::Image::new(nonzero(16), nonzero(1), fr::PixelType::U8x3);
+        let mut dst_image = fr::Image::new(nonzero(16), nonzero(1), PixelType::U8x3);
         let result = mapper.forward_map(&src_image.view(), &mut dst_image.view_mut());
         assert!(matches!(result, Err(fr::MappingError::DifferentDimensions)));
 
-        let mut dst_image = fr::Image::new(nonzero(16), nonzero(16), fr::PixelType::U8x2);
+        let mut dst_image = fr::Image::new(nonzero(16), nonzero(16), PixelType::U8x2);
         let result = mapper.forward_map(&src_image.view(), &mut dst_image.view_mut());
         assert!(matches!(
             result,
@@ -150,16 +151,16 @@ mod srgb_tests {
         let mapper = fr::create_srgb_mapper();
         let buffer: Vec<u8> = (0u8..=255).flat_map(|v| [v, v, v]).collect();
         let src_image =
-            fr::Image::from_vec_u8(nonzero(16), nonzero(16), buffer, fr::PixelType::U8x3).unwrap();
-        let src_checksum = testing::image_checksum::<3>(src_image.buffer());
+            fr::Image::from_vec_u8(nonzero(16), nonzero(16), buffer, PixelType::U8x3).unwrap();
+        let src_checksum = testing::image_checksum::<U8x3, 3>(&src_image);
         assert_eq!(src_checksum, [32640, 32640, 32640]);
 
-        let mut dst_image = fr::Image::new(nonzero(16), nonzero(16), fr::PixelType::U8x3);
+        let mut dst_image = fr::Image::new(nonzero(16), nonzero(16), PixelType::U8x3);
         mapper
             .backward_map(&src_image.view(), &mut dst_image.view_mut())
             .unwrap();
 
-        let dst_checksum = testing::image_checksum::<3>(dst_image.buffer());
+        let dst_checksum = testing::image_checksum::<U8x3, 3>(&dst_image);
         assert_eq!(dst_checksum, [44981, 44981, 44981]);
     }
 
@@ -168,13 +169,13 @@ mod srgb_tests {
         let mapper = fr::create_srgb_mapper();
         let buffer: Vec<u8> = (0u8..=255).flat_map(|v| [v, v, v]).collect();
         let src_image =
-            fr::Image::from_vec_u8(nonzero(16), nonzero(16), buffer, fr::PixelType::U8x3).unwrap();
+            fr::Image::from_vec_u8(nonzero(16), nonzero(16), buffer, PixelType::U8x3).unwrap();
 
-        let mut dst_image = fr::Image::new(nonzero(16), nonzero(1), fr::PixelType::U8x3);
+        let mut dst_image = fr::Image::new(nonzero(16), nonzero(1), PixelType::U8x3);
         let result = mapper.backward_map(&src_image.view(), &mut dst_image.view_mut());
         assert!(matches!(result, Err(fr::MappingError::DifferentDimensions)));
 
-        let mut dst_image = fr::Image::new(nonzero(16), nonzero(16), fr::PixelType::U8x2);
+        let mut dst_image = fr::Image::new(nonzero(16), nonzero(16), PixelType::U8x2);
         let result = mapper.backward_map(&src_image.view(), &mut dst_image.view_mut());
         assert!(matches!(
             result,